Calculate Variance-Covariance Matrix for a Fitted Model Object

Provides variance estimates based on the following three options

The bootstrap, Algorithm 2 in Arkhangelsky et al.

The jackknife, Algorithm 3 in Arkhangelsky et al.

Placebo, Algorithm 4 in Arkhangelsky et al.

# S3 method for synthdid_estimate
vcov(
  object,
  method = c("bootstrap", "jackknife", "placebo"),
  replications = 200,
  ...
)

Arguments

object
A synthdid model

method,
the CI method. The default is bootstrap (warning: this may be slow on large data sets, the jackknife option is the fastest, with the caveat that it is not recommended for SC).

replications,
the number of bootstrap replications

...
Additional arguments (currently ignored).

Details

The jackknife is not recommended for SC, see section 5 in Arkhangelsky et al. "placebo" is the only option that works for only one treated unit.

References

Dmitry Arkhangelsky, Susan Athey, David A. Hirshberg, Guido W. Imbens, and Stefan Wager. "Synthetic Difference in Differences". arXiv preprint arXiv:1812.09970, 2019.